refactor(workers): move worker process-tree ownership to a shared module
The ownership primitives were already generic. listWorkerProcesses matches on
whatever absolute worker path the caller passes, and classifyWorkerOwnership
takes injected liveness and PID-1 predicates. Nothing in either is specific to
embeddings; they were simply living in embedding-local.ts, which is why every
other worker has no ownership model at all and leans on a PID file instead.

Moves WorkerProcess, WorkerOwnership, classifyWorkerOwnership,
listWorkerProcesses, and the PID-1 daemon probe into
util/ml-worker-ownership.ts. The probe is renamed pid1OwnsMlWorkers, since it
answers a question about the daemon rather than about embeddings.

Behaviour is unchanged: the three moved bodies are byte-identical to their
previous versions, and the embed lifecycle tests pass against them with only
their import path updated.

Part of ATL-1349, and the enabling step for JARVIS-891.
