+++ assistant/src/persistence/embeddings/__tests__/embedding-local-lifecycle.test.ts
@@ -14,12 +14,12 @@ import { tmpdir } from "node:os";
 import { join } from "node:path";
 import { afterEach, describe, expect, test } from "bun:test";
 
-import { getEmbedWorkerPidPath } from "../../../util/platform.js";
 import {
   classifyWorkerOwnership,
   listWorkerProcesses,
-  LocalEmbeddingBackend,
-} from "../embedding-local.js";
+} from "../../../util/ml-worker-ownership.js";
+import { getEmbedWorkerPidPath } from "../../../util/platform.js";
+import { LocalEmbeddingBackend } from "../embedding-local.js";
 
 /** Reach past `private`, which is compile-time only, so tests drive real state. */
 type Internals = any;
